Sort codeModel objects
